feat: proxyFrom only supports a single event.
proxyFrom was meant to setup an SSE stream and proxy events from
an event emitter to the client. With source and destination
parameters, the events that could be proxied are limited to one. I
switched to using an event map so there can be on than one event
being proxied.